refactor: 서비스 코드 이동, 컨피그 예제 추가
